refactor: remove strict flag from AdCPSchemaValidator — unconditional strict behavior
The `strict: bool = True` parameter introduced in 65fabe72 was a hedge —
a flag that could opt out of the "No Quiet Failures" invariant. On
post-hoc review of our own diff, that hedge is wrong for this codebase:

- CLAUDE.md calls No Quiet Failures a correctness requirement, equivalent
  to type safety. A flag that disables a correctness requirement is a
  code smell.
- Every existing structural guard in the project ships with zero
  configuration knobs. We should match that posture.
- Grep confirms zero callers pass `strict=False` anywhere in the repo —
  the opt-out was pure optionality nobody used.
- "Strict" is a semantically overloaded term (validation mode? schema
  matching? type coercion?); removing it eliminates future ambiguity.

Changes to tests/e2e/adcp_schema_validator.py:

- Remove `strict` parameter from AdCPSchemaValidator.__init__; drop
  `self.strict` attribute.
- _resolve_adcp_schema_ref: unconditionally raises SchemaResolutionError
  on cache miss or corrupt file. Error message embeds
  `make schemas-refresh` remediation hint — a missing-schema error is
  useless without it.
- _resolve_http_schema_ref: unconditionally raises on unknown host.
  Delegates adcontextprotocol.org URLs to the AdCP resolver.
- _preload_schema_references: bare propagation. Removed the redundant
  try/except that existed solely to gate the raise on self.strict.
- Delete preload_schemas() helper + __main__ block. Zero external
  callers (verified via grep); superseded by
  scripts/ops/refresh_adcp_schemas.py's full transitive-closure fetch.
  The 6 `except Exception as e: print(f"⚠ …")` swallows inside
  preload_schemas were the exact antipattern #1213 fixes, in a
  different code path; deletion eliminates them instead of migrating
  them. Also removes the now-unused `asyncio` import.
- Rewrite SchemaResolutionError docstring: drop "strict mode" framing,
  describe the unconditional contract and the refresh remediation.

Renames tests/unit/test_adcp_schema_validator_strictness.py →
test_adcp_schema_validator_ref_resolution.py. "Strictness" described a
flag that no longer exists; "ref_resolution" describes the actual
contract being tested. Drops the 5 lenient-mode tests (behavior no
longer exists), the 2-test TestStrictModeDefaults class, and removes
`strict=True` kwargs from remaining test constructors. Keeps 7 tests
covering: SchemaResolutionError shape, missing-cache-file raises,
corrupt-cache-file raises, unknown-HTTP-host raises, adcontextprotocol
delegation, and cache-hit returns. Adds `make schemas-refresh` hint
assertions on the raise messages.

Updates docs/development/structural-guards.md: rewrite the
"strict=True mode" language in the AdCP Schema Cache Completeness Guard
section to reflect unconditional raise on cache miss.

Verified: 4271 unit tests pass (was 4276 — 5 lenient-mode tests removed
as expected), 17 e2e schema tests pass under unconditional strict (zero
drift surfaced).
